refactor: utils -> lib
