+++ src/entities/offer-card/lib/mocks.ts
